requirements.txt
# Google 表格与 Drive
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# 下载 Instagram 视频
yt-dlp>=2024.1.0

# 环境变量
python-dotenv>=1.0.0
